requirements.txt
pandas
langchain-core
langchain-ollama
langchain-chroma
langchain-community
